Api validation

+++ src/api/admin/places.js
@@ -3,6 +3,16 @@ const Place = require("../../model/Place");
 const Review = require('../../model/Review');
 const ImageModel = require('../../model/Image');
 
+
+const { param, validationResult } = require("express-validator");
+const mongoose = require('mongoose')
+
+const isValidObjectId = param("id").custom((value) => {
+  const isValid = mongoose.Types.ObjectId.isValid(value);
+  if (!isValid) throw new Error("Invalid ID");
+  return isValid;
+});
+
 router.get("/", async (req, res, next) => {
   try {
     const query = Place.find({});
@@ -29,8 +39,13 @@ router.get("/flagged", async (req, res, next) => {
   }
 });
 
-router.get("/:id", async (req, res) => {
+router.get("/:id", isValidObjectId, async (req, res) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+
     const place = await Place.findById(req.params.id).populate(
       "user",
       "-password"
@@ -42,8 +57,13 @@ router.get("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", isValidObjectId, async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+
     const place = await Place.findById(req.params.id);
     if (!place) res.sendStatus(404);
     await place.delete();
@@ -53,8 +73,13 @@ router.delete("/:id", async (req, res, next) => {
   }
 });
 
-router.get("/:id/reviews", async (req, res, next) => {
+router.get("/:id/reviews", isValidObjectId, async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+
     const reviews = await Review.find({
       place: req.params.id,
     }).populate("user", "-password");
@@ -64,8 +89,13 @@ router.get("/:id/reviews", async (req, res, next) => {
   }
 });
 
-router.get("/:id/images", async (req, res, next) => {
+router.get("/:id/images", isValidObjectId, async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+    
     const images = await ImageModel.find({
       place: req.params.id,
     }).populate("user", "-password");
